Fix stream parameters in asynchronous methods.
Within gpuarray.set_async, the stream keyword
argument in the call to gpuarray.set was
always set to None.

  return self.set(ary, async=True, stream=None)

This keyword argument is now always set to value
supplied in the arguments to gpuarray.set_async.

  return self.set(ary, async=True, stream=stream)

Within _memcpy_discontig, Memcpy2D.__call__
was being called for asynchronous cases as follows:

  copy(stream=stream)

However, this does not match the C++ function
definition. Instead, they are now called as follows:

  copy(stream)